[HOUSEKEEPING] move functions and remove files
