+++ src/lib/azdoClient.test.ts
@@ -10,12 +10,8 @@ import {
   getTaskAgentApi,
   getWebApi,
   invalidateWebApi,
-  repositoryHasFile,
-  validateRepository,
 } from "./azdoClient";
 import * as azdoClient from "./azdoClient";
-import { AzureDevOpsOpts } from "./git";
-import * as azure from "./git/azure";
 
 describe("AzDo Pipeline utility functions", () => {
   test("azdo url is well formed", () => {
@@ -141,109 +137,5 @@ describe("test getBuildApi function", () => {
     mockConfig(); // empty config. still work because API client is cached
     const again = await getBuildApi();
     expect(again).toBeDefined();
-  });
-});
-
-describe("validateRepository", () => {
-  test("repository exists", async () => {
-    const getRepositoryFunc = jest.spyOn(azure, "GitAPI");
-    getRepositoryFunc.mockReturnValueOnce(
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      Promise.resolve({ getRepository: () => ({ id: "3839fjfkj" }) } as any)
-    );
-    const getItemFunc = jest.spyOn(azure, "GitAPI");
-    getItemFunc.mockReturnValueOnce(
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      Promise.resolve({ getItem: () => ({ commitId: "3839fjfkj" }) } as any)
-    );
-    const accessOpts: AzureDevOpsOpts = {
-      orgName: "testOrg",
-      personalAccessToken: "mytoken",
-      project: "testProject",
-    };
-
-    await expect(
-      validateRepository(
-        "my-project",
-        "myFile",
-        "master",
-        "my-repo",
-        accessOpts
-      )
-    ).resolves.not.toThrow();
-  });
-  test("repository does not exist", async () => {
-    const createPullRequestFunc = jest.spyOn(azure, "GitAPI");
-    createPullRequestFunc.mockReturnValueOnce(
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      Promise.resolve({ getRepository: () => null } as any)
-    );
-    const accessOpts: AzureDevOpsOpts = {
-      orgName: "testOrg",
-      personalAccessToken: "mytoken",
-      project: "testProject",
-    };
-    await expect(
-      validateRepository(
-        "my-project",
-        "myFile",
-        "master",
-        "my-repo",
-        accessOpts
-      )
-    ).rejects.toThrow();
-  });
-});
-
-describe("repositoryHasFile", () => {
-  test("repository contains the given file", async () => {
-    const createPullRequestFunc = jest.spyOn(azure, "GitAPI");
-    createPullRequestFunc.mockReturnValueOnce(
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      Promise.resolve({ getItem: () => ({ commitId: "3839fjfkj" }) } as any)
-    );
-    const accessOpts: AzureDevOpsOpts = {
-      orgName: "testOrg",
-      personalAccessToken: "mytoken",
-      project: "testProject",
-    };
-    let hasError = false;
-
-    try {
-      await repositoryHasFile(
-        "testFile.txt",
-        "master",
-        "test-repo",
-        accessOpts
-      );
-    } catch (err) {
-      hasError = true;
-    }
-    expect(hasError).toBe(false);
-  });
-  test("repository does not contain the given file", async () => {
-    const createPullRequestFunc = jest.spyOn(azure, "GitAPI");
-    createPullRequestFunc.mockReturnValueOnce(
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      Promise.resolve({ getItem: () => null } as any)
-    );
-    const accessOpts: AzureDevOpsOpts = {
-      orgName: "testOrg",
-      personalAccessToken: "mytoken",
-      project: "testProject",
-    };
-    let hasError = false;
-
-    try {
-      await repositoryHasFile(
-        "testFile2.txt",
-        "master",
-        "test-repo",
-        accessOpts
-      );
-    } catch (err) {
-      hasError = true;
-    }
-    expect(hasError).toBe(true);
   });
 });
